feat: changed the name from v2 -> v1 and v3 -> v2
