+++ pkg/engine/defaults.go
@@ -113,8 +113,8 @@ import (
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/campayn"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/cannyio"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/capsulecrm"
+	captainDataV1 "github.com/trufflesecurity/trufflehog/v3/pkg/detectors/captaindata/v1"
 	captainDataV2 "github.com/trufflesecurity/trufflehog/v3/pkg/detectors/captaindata/v2"
-	captainDataV3 "github.com/trufflesecurity/trufflehog/v3/pkg/detectors/captaindata/v3"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/carboninterface"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/cashboard"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors/caspio"
@@ -1642,8 +1642,8 @@ func DefaultDetectors() []detectors.Detector {
 		saladcloudapikey.Scanner{},
 		boxoauth.Scanner{},
 		apimetrics.Scanner{},
+		captainDataV1.Scanner{},
 		captainDataV2.Scanner{},
-		captainDataV3.Scanner{},
 	}
 
 	// Automatically initialize all detectors that implement
